dropwatch: Add some filtering support
Some _very_ rudimentary filtering support for dropwatch. Adds the keyword filter
to the interactive shell for dropwatch.  With this commit you can specify this
command:
dropwatch> filter <address>
where address is the hex value that is output for a given location (including
the leading 0x) while monitoring drops.  Adding a location filter will cause
drops at that location to no longer be displayed.

Signed-off-by: Neil Horman <nhorman@redhat.com>